TST: re-arrange tests
So that we can easily re-use the tests else where
