+++ filestore/test/test_fs_mutate.py
@@ -8,7 +8,6 @@ import uuid
 import numpy as np
 
 from filestore.handlers_base import HandlerBase
-from .utils import insert_syn_data_with_resource
 
 
 def _verify_shifted_resource(last_res, new_res):
@@ -233,20 +232,3 @@ def test_temporary_root(fs_v1):
         path = fs.retrieve(dm['datum_id'])
 
     assert path == os.path.join('baz', 'foo')
-
-
-def test_read_old_in_new_resource(fs_v01):
-    fs0, fs1 = fs_v01
-    shape = [25, 32]
-    # save data using old schema
-    mod_ids, res = insert_syn_data_with_resource(fs0, 'syn-mod',
-                                                 shape, 10)
-
-    resource = fs0.resource_given_uid(res)
-    assert resource == res
-    assert res == fs1.resource_given_uid(res)
-
-    for j, d_id in enumerate(mod_ids):
-        # get back using new schema
-        d_res = fs1.resource_given_eid(d_id)
-        assert d_res == resource
